+++ gui.py
@@ -16,7 +16,7 @@ Usage:
     python gui.py
 """
 
-__version__ = "1.2.5"
+__version__ = "1.2.7"
 __author__ = "Finlay Crawley"
 
 import sys